created ListenerUtils and moved common functions in it.
