+++ epoxy-adapter/src/main/java/com/airbnb/epoxy/WrappedEpoxyModelCheckedChangeListener.java
@@ -1,13 +1,9 @@
 package com.airbnb.epoxy;
 
-import android.view.View;
-import android.view.ViewParent;
 import android.widget.CompoundButton;
 import android.widget.CompoundButton.OnCheckedChangeListener;
 
-import androidx.annotation.Nullable;
 import androidx.recyclerview.widget.RecyclerView;
-import androidx.recyclerview.widget.RecyclerView.ViewHolder;
 
 /**
  * Used in the generated models to transform normal checked change listener to model
@@ -30,7 +26,7 @@ public class WrappedEpoxyModelCheckedChangeListener<T extends EpoxyModel<?>, V>
 
   @Override
   public void onCheckedChanged(CompoundButton button, boolean isChecked) {
-    EpoxyViewHolder epoxyHolder = getEpoxyHolderForChildView(button);
+    EpoxyViewHolder epoxyHolder = ListenersUtils.getEpoxyHolderForChildView(button);
     if (epoxyHolder == null) {
       throw new IllegalStateException("Could not find RecyclerView holder for clicked view");
     }
@@ -41,43 +37,6 @@ public class WrappedEpoxyModelCheckedChangeListener<T extends EpoxyModel<?>, V>
           .onChecked((T) epoxyHolder.getModel(), (V) epoxyHolder.objectToBind(), button,
           isChecked, adapterPosition);
     }
-  }
-
-  @Nullable
-  private static EpoxyViewHolder getEpoxyHolderForChildView(View v) {
-    RecyclerView recyclerView = findParentRecyclerView(v);
-    if (recyclerView == null) {
-      return null;
-    }
-
-    ViewHolder viewHolder = recyclerView.findContainingViewHolder(v);
-    if (viewHolder == null) {
-      return null;
-    }
-
-    if (!(viewHolder instanceof EpoxyViewHolder)) {
-      return null;
-    }
-
-    return (EpoxyViewHolder) viewHolder;
-  }
-
-  @Nullable
-  private static RecyclerView findParentRecyclerView(@Nullable View v) {
-    if (v == null) {
-      return null;
-    }
-
-    ViewParent parent = v.getParent();
-    if (parent instanceof RecyclerView) {
-      return (RecyclerView) parent;
-    }
-
-    if (parent instanceof View) {
-      return findParentRecyclerView((View) parent);
-    }
-
-    return null;
   }
 
   @Override
