+++ epoxy-adapter/src/main/java/com/airbnb/epoxy/WrappedEpoxyModelClickListener.java
@@ -3,11 +3,8 @@ package com.airbnb.epoxy;
 import android.view.View;
 import android.view.View.OnClickListener;
 import android.view.View.OnLongClickListener;
-import android.view.ViewParent;
 
-import androidx.annotation.Nullable;
 import androidx.recyclerview.widget.RecyclerView;
-import androidx.recyclerview.widget.RecyclerView.ViewHolder;
 
 /**
  * Used in the generated models to transform normal view click listeners to model click
@@ -40,7 +37,7 @@ public class WrappedEpoxyModelClickListener<T extends EpoxyModel<?>, V>
 
   @Override
   public void onClick(View v) {
-    EpoxyViewHolder epoxyHolder = getEpoxyHolderForChildView(v);
+    EpoxyViewHolder epoxyHolder = ListenersUtils.getEpoxyHolderForChildView(v);
     if (epoxyHolder == null) {
       throw new IllegalStateException("Could not find RecyclerView holder for clicked view");
     }
@@ -55,7 +52,7 @@ public class WrappedEpoxyModelClickListener<T extends EpoxyModel<?>, V>
 
   @Override
   public boolean onLongClick(View v) {
-    EpoxyViewHolder epoxyHolder = getEpoxyHolderForChildView(v);
+    EpoxyViewHolder epoxyHolder = ListenersUtils.getEpoxyHolderForChildView(v);
     if (epoxyHolder == null) {
       throw new IllegalStateException("Could not find RecyclerView holder for clicked view");
     }
@@ -69,43 +66,6 @@ public class WrappedEpoxyModelClickListener<T extends EpoxyModel<?>, V>
     }
 
     return false;
-  }
-
-  @Nullable
-  private static EpoxyViewHolder getEpoxyHolderForChildView(View v) {
-    RecyclerView recyclerView = findParentRecyclerView(v);
-    if (recyclerView == null) {
-      return null;
-    }
-
-    ViewHolder viewHolder = recyclerView.findContainingViewHolder(v);
-    if (viewHolder == null) {
-      return null;
-    }
-
-    if (!(viewHolder instanceof EpoxyViewHolder)) {
-      return null;
-    }
-
-    return (EpoxyViewHolder) viewHolder;
-  }
-
-  @Nullable
-  private static RecyclerView findParentRecyclerView(@Nullable View v) {
-    if (v == null) {
-      return null;
-    }
-
-    ViewParent parent = v.getParent();
-    if (parent instanceof RecyclerView) {
-      return (RecyclerView) parent;
-    }
-
-    if (parent instanceof View) {
-      return findParentRecyclerView((View) parent);
-    }
-
-    return null;
   }
 
   @Override
